update for World and Query classes for future refactoring of archetype class.

+++ GameComponents_ECS/Systems/Archetype/Archetype.cs
@@ -6,7 +6,7 @@ public partial record Archetype
 {
     private readonly string _collectedTypes;
     
-    internal short _nextPosition;
+    internal int _nextPosition;
     internal readonly sbyte[] _indexMap;
     internal readonly BitArray _bits;
